# perf(bash-v2): speed up filtering entries with descriptions