@marckhouzam 2022-05-04
With @scop's multiple optimizations for the bash completion V2 script, we have gone from

bash V1 1000 comps (no descriptions): **0.072** seconds (no optimizations done nor needed for V1)
bash V2 1000 comps with no descriptions: from **3.112** seconds to **0.052** seconds
bash V2 1000 comps with descriptions: from **3.479** seconds to **0.241** seconds

Amazing! 💥  💥 💥 
I think we're good 😆