@scop 2022-05-04
Turns out there was room to sneak in a bug here after all :stuck_out_tongue:.  Fix in #1691.